[hil] added counter in Hil objects

+++ src/Main/HIL/HIL.h
@@ -115,7 +115,7 @@ public:
 
 private:
     ActuatorData updateActuatorData();
-    uint8_t counter = 0;
+    uint64_t counter = 0;
 };
 
 }  // namespace Main
