+++ src/Motor/HIL/HIL.h
@@ -74,6 +74,6 @@ public:
 
 private:
     ActuatorData updateActuatorData();
-    uint8_t counter = 0;
+    uint64_t counter = 0;
 };
 }  // namespace Motor
